requirements.txt
streamlit
pillow
tensorflow
keras
redis